+++ lib/src/auth/auth.dart
@@ -5,12 +5,13 @@ import 'objects.dart';
 
 export 'auth_code_controller.dart';
 export 'auth_controller.dart';
-export 'authorization_server.dart';
 export 'authorization_parser.dart';
+export 'authorization_server.dart';
 export 'authorizer.dart';
 export 'exceptions.dart';
 export 'objects.dart';
 export 'protocols.dart';
+export 'validator.dart';
 
 /// Exposes static utility methods for password, salt and API credential generation.
 ///