+++ lib/src/auth/authorizer.dart
@@ -19,119 +19,111 @@ enum AuthStrategy {
 
 /// A [Controller] that validates the Authorization header of a request.
 ///
-/// An instance of [Authorizer] will validate a [Request] given a [strategy] and a [validator]. [validator] is typically the instance
-/// of [AuthServer] in an application.
+/// An instance of this type will validate that the authorization information in an Authorization header is sufficient to access
+/// the next controller in the channel.
 ///
-/// If the [Request] is unauthorized (as determined by the [validator]), it will respond with the appropriate status code and prevent
-/// further request processing. If the [Request] is valid, an [Authorization] will be added
-/// to the [Request] and the request will be delivered to this instance's [nextController]. Usage:
+/// For each request, this controller parses the authorization header, validates it with an [AuthValidator] and then create an [Authorization] object
+/// if successful. The [Request] keeps a reference to this [Authorization] and is then sent to the next controller in the channel.
+///
+/// If either parsing or validation fails, a 401 Unauthorized response is sent and the [Request] is removed from the channel.
+///
+/// Parsing occurs according to [parser]. The resulting value (e.g., username and password) is sent to [validator].
+/// [validator] verifies this value (e.g., lookup a user in the database and verify their password matches).
+///
+/// Usage:
 ///
 ///         router
-///           .route("/protectedroute")
+///           .route("/protected-route")
 ///           .pipe(new Authorizer.bearer(authServer))
 ///           .generate(() => new ProtectedResourceController());
 class Authorizer extends Controller {
   /// Creates an instance of [Authorizer].
   ///
-  /// The default strategy is [AuthStrategy.bearer].
-  Authorizer(this.validator,
-      {this.strategy: AuthStrategy.bearer, List<String> scopes}) {
-    this.scopes = scopes;
+  /// Use this constructor to provide custom [AuthorizationParser]s.
+  ///
+  /// By default, this instance will parse bearer tokens from the authorization header, e.g.:
+  ///
+  ///         Authorization: Bearer ap9ijlarlkz8jIOa9laweo
+  ///
+  /// If [scopes] is provided, the authorization granted must have access to *all* scopes according to [validator].
+  Authorizer(this.validator, {this.parser: const AuthorizationBearerParser(), List<String> scopes})
+      : this.scopes = scopes?.map((s) => new AuthScope(s))?.toList() {
     policy = null;
   }
 
-  /// Creates an instance of [Authorizer] using [AuthStrategy.basic].
-  Authorizer.basic(AuthValidator validator)
-      : this(validator, strategy: AuthStrategy.basic);
-
-  /// Creates an instance of [Authorizer] using [AuthStrategy.bearer].
+  /// Creates an instance of [Authorizer] with Basic Authentication parsing.
   ///
-  /// Optionally allows the setting of [Authorizer.scopes].
+  /// Parses a username and password from the request's Basic Authentication data in the Authorization header, e.g.:
+  ///
+  ///         Authorization: Basic base64(username:password)
+  Authorizer.basic(AuthValidator validator) : this(validator, parser: const AuthorizationBasicParser());
+
+  /// Creates an instance of [Authorizer] with Bearer token parsing.
+  ///
+  /// Parses a bearer token from the request's Authorization header, e.g.
+  ///
+  ///         Authorization: Bearer ap9ijlarlkz8jIOa9laweo
+  ///
+  /// If [scopes] is provided, the bearer token must have access to *all* scopes according to [validator].
   Authorizer.bearer(AuthValidator validator, {List<String> scopes})
-      : this(validator, strategy: AuthStrategy.bearer, scopes: scopes);
+      : this(validator, parser: const AuthorizationBearerParser(), scopes: scopes);
 
   /// The validating authorization object.
   ///
   /// This object will check credentials parsed from the Authorization header and produce an
-  /// [Authorization] instance representing the authorization the credentials have. It may also
+  /// [Authorization] object representing the authorization the credentials have. It may also
   /// reject a request. This is typically an instance of [AuthServer].
-  AuthValidator validator;
+  final AuthValidator validator;
 
-  /// The list of scopes this instance requires.
+  /// The list of required scopes.
   ///
-  /// A bearer token must have access to all of the scopes in this list in order to pass
-  /// through to the [nextController].
-  List<String> get scopes => _scopes?.map((s) => s.scopeString)?.toList();
-  set scopes(List<String> scopes) {
-    _scopes = scopes?.map((s) => new AuthScope(s))?.toList();
-  }
-  List<AuthScope> _scopes;
+  /// If [validator] grants scope-limited authorizations (e.g., OAuth2 bearer tokens), the authorization
+  /// provided by the request's header must have access to all [scopes] in order to move on to the next controller.
+  ///
+  /// This property is set with a list of scope strings in a constructor. Each scope string is parsed into
+  /// an [AuthScope] and added to this list.
+  final List<AuthScope> scopes;
 
-  /// The [AuthStrategy] for authorizing a request.
+  /// Parses the Authorization header.
   ///
-  /// This property determines which [AuthValidator] method is invoked on [validator].
-  AuthStrategy strategy;
+  /// The parser determines how to interpret the data in the Authorization header. Concrete subclasses
+  /// are [AuthorizationBasicParser] and [AuthorizationBearerParser].
+  ///
+  /// Once parsed, the parsed value is validated by [validator].
+  final AuthorizationParser parser;
 
   @override
-  FutureOr<RequestOrResponse> handle(Request req) {
-    var header = req.raw.headers.value(HttpHeaders.AUTHORIZATION);
-    if (header == null) {
+  FutureOr<RequestOrResponse> handle(Request req) async {
+    var authData = req.raw.headers.value(HttpHeaders.AUTHORIZATION);
+    if (authData == null) {
       return new Response.unauthorized();
     }
 
-    switch (strategy) {
-      case AuthStrategy.bearer: return _processBearerHeader(req, header);
-      case AuthStrategy.basic: return _processBasicHeader(req, header);
-      default: return new Response.serverError();
-    }
-  }
-
-  Future<RequestOrResponse> _processBearerHeader(
-      Request request, String headerValue) async {
-    String bearerToken;
+    var value;
     try {
-      bearerToken = AuthorizationBearerParser.parse(headerValue);
+      value = parser.parse(authData);
     } on AuthorizationParserException catch (e) {
       return _responseFromParseException(e);
     }
 
-    var authorization = await validator.fromBearerToken(bearerToken, scopesRequired: _scopes);
-    if (authorization == null) {
+    req.authorization = await validator.validate(parser, value, requiredScope: scopes);
+
+    if (req.authorization == null) {
       return new Response.unauthorized();
     }
 
-    request.authorization = authorization;
-    return request;
-  }
-
-  Future<RequestOrResponse> _processBasicHeader(
-      Request request, String headerValue) async {
-    AuthBasicCredentials elements;
-    try {
-      elements = AuthorizationBasicParser.parse(headerValue);
-    } on AuthorizationParserException catch (e) {
-      return _responseFromParseException(e);
-    }
-
-    var authorization = await validator.fromBasicCredentials(elements);
-    if (authorization == null) {
-      return new Response.unauthorized();
-    }
-
-    request.authorization = authorization;
-    return request;
+    return req;
   }
 
   Response _responseFromParseException(AuthorizationParserException e) {
     switch (e.reason) {
       case AuthorizationParserExceptionReason.malformed:
-        return new Response.badRequest(
-            body: {"error": "invalid_authorization_header"});
+        return new Response.badRequest(body: {"error": "invalid_authorization_header"});
       case AuthorizationParserExceptionReason.missing:
         return new Response.unauthorized();
       default:
         return new Response.serverError();
-    }    
+    }
   }
 
 
@@ -147,39 +139,4 @@ class Authorizer extends Controller {
 
     return operations;
   }
-}
-
-/// Instances that implement this type can be used by an [Authorizer] to determine authorization of a request.
-///
-/// When an [Authorizer] processes a [Request], it invokes methods from this type to determine the [Authorization] from the Authorization
-/// header of the [Request]. [AuthServer] implements this interface.
-abstract class AuthValidator {
-  /// Returns an [Authorization] from basic credentials.
-  ///
-  /// This method must either return a [Future] that yields an [Authorization] or return null.
-  /// If this method returns null, the invoking [Authorizer] will disallow further
-  /// request handling and immediately return a 401 status code. If this method returns an
-  /// [Authorization], it will be set as the [Request.authorization] and request handling
-  /// will continue to the [Authorizer.nextController].
-  FutureOr<Authorization> fromBasicCredentials(
-      AuthBasicCredentials usernameAndPassword);
-
-  /// Returns an [Authorization] from a bearer token.
-  ///
-  /// This method must either return a [Future] that yields an [Authorization] or return null.
-  /// If this method returns null, the invoking [Authorizer] will disallow further
-  /// request handling and immediately return a 401 status code. If this method returns an
-  /// [Authorization], it will be set as the [Request.authorization] and request handling
-  /// will continue to the [Authorizer.nextController].
-  ///
-  /// [scopesRequired] is the list of scopes established when the [Authorizer]
-  /// is created. Implementors of this method must verify the bearer token has access to [scopesRequired].
-  ///
-  /// If [scopesRequired] is null, an implementor may make its own determination about whether
-  /// the token results in an [Authorization]. By default, [AuthServer] - the primary implementor of this type -
-  /// will allow access, assuming that 'null scope' means 'any scope'.
-  FutureOr<Authorization> fromBearerToken(
-      String bearerToken, {List<AuthScope> scopesRequired});
-
-  List<APISecurityRequirement> documentSecurityRequirements(AuthStrategy strategy, {List<String> scopes}) => [];
 }